incoming/crampus/repo-agent/pyproject.toml
[project]
name = "repo-agent"
version = "0.1.0"
description = "Run a configurable LLM-powered code tool in Docker against a repo using markdown instructions, then push a branch and create a PR."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12.3",
    "python-frontmatter>=1.1.0",
]

[project.scripts]
repo-agent = "repo_agent.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
index-url = "https://pypi.org/simple"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

[tool.coverage.run]
source = ["src/repo_agent"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
